Moving components into separate folder
